Build MPS3 AN547 with GCC
Since this code was written, GCC now compiles for the M55. From what I
can tell (by diffing scatter files), the an524 and the an547 have very
similar code layout. So the linker scripts used are from the an524.
This also inclues a note in the documentation abouth the required gcc
version.

Change-Id: I365ad3f934956a46756293694a44cca08ea25898
Signed-off-by: Jimmy Brisson <jimmy.brisson@linaro.org>